fix: Returned rag.py to original state and created new rag_obs_eval so that we can demonstrate a before and after obs_eval state.
